NLP-Project/yt-summarize-sentiment-analyze/requirements.txt
# YouTube NLP Analyzer - Dependencies

# YouTube transcript extraction
youtube-transcript-api==0.6.2

# OpenAI API for LLM-based summarization and sentiment analysis
openai==1.12.0

# Additional utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For alternative summarization (if OpenAI is not used)
# transformers==4.36.0
# torch==2.1.0
# sentencepiece==0.1.99

# Development dependencies (optional)
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0
